Remaining Bug Fixes
Remaining Bug Fixes

+++ advanced-user-registration-and-management/lr-social-invite/includes/helpers/ajax.php
@@ -3,27 +3,27 @@
 class Ajax_Social_Invite_Helper {
 
     public function __construct() {
-        add_action('wp_ajax_social_invite_get_contacts', array($this, 'get_contacts'), 1);
-        add_action('wp_ajax_social_invite_search_contacts', array($this, 'search_contacts'), 1);
-        add_action('wp_ajax_social_invite_is_token_valid', array($this, 'is_token_valid'), 1);
-        add_action('wp_ajax_social_invite_get_provider_token', array($this, 'get_provider_token'), 1);
-        add_action('wp_ajax_social_invite_update_provider_token', array($this, 'update_provider_token'), 1);
-        add_action('wp_ajax_social_invite_send_message', array($this, 'send_message'), 1);
+        add_action( 'wp_ajax_social_invite_get_contacts', array( $this, 'get_contacts' ), 1 );
+        add_action( 'wp_ajax_social_invite_search_contacts', array( $this, 'search_contacts' ), 1 );
+        add_action( 'wp_ajax_social_invite_is_token_valid', array( $this, 'is_token_valid' ), 1 );
+        add_action( 'wp_ajax_social_invite_get_provider_token', array( $this, 'get_provider_token' ), 1);
+        add_action( 'wp_ajax_social_invite_update_provider_token', array( $this, 'update_provider_token' ), 1 );
+        add_action( 'wp_ajax_social_invite_send_message', array( $this, 'send_message' ), 1 );
     }
 
     public function get_provider_token() {
         global $wpdb;
 
-        $provider = isset($_POST['provider']) ? $_POST['provider'] : '';
+        $provider = isset( $_POST['provider'] ) ? $_POST['provider'] : '';
 
-        if (!empty($provider)) {
+        if ( ! empty( $provider ) ) {
             // If the user is logged in
             $user = wp_get_current_user();
             $results = $wpdb->get_results(
-                    $wpdb->prepare("SELECT * FROM " . $wpdb->base_prefix . "lr_social_invite_tokens WHERE user_id = %d AND provider = %s", $user->ID, $_POST['provider'])
+                    $wpdb->prepare( "SELECT * FROM " . $wpdb->base_prefix . "lr_social_invite_tokens WHERE user_id = %d AND provider = %s", $user->ID, $_POST['provider'] )
             );
 
-            echo json_encode($results);
+            echo json_encode( $results );
         }
 
         die();
@@ -35,43 +35,42 @@ class Ajax_Social_Invite_Helper {
         // If the user is logged in
         $user = wp_get_current_user();
 
-        $provider = isset($_POST['provider']) ? $_POST['provider'] : '';
-        $token = isset($_POST['token']) ? $_POST['token'] : '';
+        $provider = isset( $_POST['provider'] ) ? $_POST['provider'] : '';
+        $token = isset( $_POST['token'] ) ? $_POST['token'] : '';
 
-        if (!empty($provider) && !empty($token)) {
+        if ( ! empty( $provider ) && ! empty( $token ) ) {
 
             $date = new DateTime();
-            $datestring = $date->format('Y-m-d H:i:s');
+            $datestring = $date->format( 'Y-m-d H:i:s' );
 
             $data = array();
 
             $data['user_id'] = $user->ID;
             $data['provider'] = $provider;
 
-            $delete_response = $wpdb->delete( $wpdb->base_prefix . 'lr_social_invite_tokens', $data);
+            $delete_response = $wpdb->delete( $wpdb->base_prefix . 'lr_social_invite_tokens', $data );
 
             $data['token'] = $token;
             $data['creationdatetime'] = $datestring;
-            $insert_response = $wpdb->insert( $wpdb->base_prefix . 'lr_social_invite_tokens', $data);
+            $insert_response = $wpdb->insert( $wpdb->base_prefix . 'lr_social_invite_tokens', $data );
 
-            echo json_encode($array = array("deleteResponse" => $delete_response, "insertResponse" => $insert_response));
+            echo json_encode( $array = array("deleteResponse" => $delete_response, "insertResponse" => $insert_response ) );
         }
-
         die();
     }
 
     public function is_token_valid() {
         global $loginradius_api_settings;
 
-        $token = isset($_POST['token']) ? $_POST['token'] : '';
+        $token = isset( $_POST['token'] ) ? $_POST['token'] : '';
 
-        if (!empty($token)) {
+        if ( ! empty( $token ) ) {
             $loginRadiusObject = new LoginRadius();
             $Response = $loginRadiusObject->loginradius_fetch_access_token( $loginradius_api_settings['LoginRadius_secret'], $token, true );
-            if (isset($Response->access_token) && $Response->access_token != '') {
-                echo json_encode($Response->access_token);
+            if ( isset( $Response->access_token ) && $Response->access_token != '' ) {
+                echo json_encode( $Response->access_token );
             } else {
-                echo json_encode("false");
+                echo json_encode( "false" );
             }
         }
 
@@ -83,32 +82,43 @@ class Ajax_Social_Invite_Helper {
         $loginRadiusObject = new LoginRadius();
 
         $accessToken = $_POST['token'];
-        $contacts = $loginRadiusObject->loginradius_get_contacts($accessToken);
+        
+        $cursor = '0';
+        
+        $contacts = array();
+  
+        try {
+            $contacts = $loginRadiusObject->loginradius_get_contacts( $accessToken, $cursor ); 
+        } catch( LoginRadiusException $e ) {
+            $contacts = null;
+            error_log( $e->errorResponse->message );
+            die( json_encode( array( 'error' => $e->errorResponse->message ) ) );
+        }
 
         // If the user is logged in
         $user = wp_get_current_user();
 
-        if (!$user->exists()) {
-            return;
+        if ( ! $user->exists() || ! isset( $contacts ) ) {
+            die( json_encode( array( 'error' => 'WordPress user is not logged in or no contacts were received' ) ) );
         }
 
-        if (!empty($_POST['provider']) && !empty($_POST['token'])) {
+        if ( ! empty( $_POST['provider'] ) && ! empty( $_POST['token'] ) ) {
 
             $date = new DateTime();
-            $datestring = $date->format('Y-m-d H:i:s');
+            $datestring = $date->format( 'Y-m-d H:i:s' );
 
             $data = array();
             $data['user_id'] = $user->ID;
             $data['provider'] = $_POST['provider'];
-            $wpdb->delete($wpdb->base_prefix . 'lr_social_invite_contacts', $data);
-            $wpdb->delete($wpdb->base_prefix . 'lr_social_invite_tokens', $data);
+            $wpdb->delete( $wpdb->base_prefix . 'lr_social_invite_contacts', $data );
+            $wpdb->delete( $wpdb->base_prefix . 'lr_social_invite_tokens', $data );
 
             $data['token'] = $_POST['token'];
             $data['creationdatetime'] = $datestring;
-            $wpdb->insert($wpdb->base_prefix . 'lr_social_invite_tokens', $data);
+            $wpdb->insert( $wpdb->base_prefix . 'lr_social_invite_tokens', $data );
 
 
-            for ($i = 0; $i < count($contacts->Data); $i++) {
+            for ( $i = 0; $i < count( $contacts->Data ); $i++ ) {
 
                 // create array to insert data
 
@@ -128,24 +138,25 @@ class Ajax_Social_Invite_Helper {
                 $data['gender'] = $contacts->Data[$i]->Gender;
 
 
-                if ($data['name'] == "") {
-                    $emailname = explode('@', $data['email']);
+                if ( $data['name'] == "" ) {
+                    $emailname = explode( '@', $data['email'] );
                     $data['name'] = $emailname[0];
                 }
-                $data['name'] = ucfirst($data['name']);
+                $data['name'] = ucfirst( $data['name'] );
 
                 // Set date.
                 $date = $contacts->Data[$i]->DateOfBirth;
-                if ($date != NULL) {
-                    $time = strtotime($date);
-                    $data['dob'] = date('Y-m-d', $time);
+                if ( $date != NULL ) {
+                    $time = strtotime( $date );
+                    $data['dob'] = date( 'Y-m-d', $time );
                 }
 
-                if ($data['name'] != 'Private private') {
-                    $wpdb->insert($wpdb->base_prefix . 'lr_social_invite_contacts', $data);
+                if ( $data['name'] != 'Private private' ) {
+                    $wpdb->insert( $wpdb->base_prefix . 'lr_social_invite_contacts', $data );
                 }
             }
         }
+
         die();
     }
 
@@ -165,16 +176,16 @@ class Ajax_Social_Invite_Helper {
         // If the user is logged in
         $user = wp_get_current_user();
 
-        $query = $wpdb->prepare("SELECT * FROM " . $wpdb->base_prefix . "lr_social_invite_contacts WHERE user_id = %d " . $sort, $user->ID);
+        $query = $wpdb->prepare( "SELECT * FROM " . $wpdb->base_prefix . "lr_social_invite_contacts WHERE user_id = %d " . $sort, $user->ID );
 
-        if (!empty($search)) {
+        if ( ! empty( $search ) ) {
             $search_query = "%" . $search . "%";
             $query = $wpdb->prepare("SELECT * FROM " . $wpdb->base_prefix . "lr_social_invite_contacts WHERE user_id = %d AND name LIKE %s " . $sort, $user->ID, $search_query);
         }
 
-        $results = $wpdb->get_results($query);
+        $results = $wpdb->get_results( $query );
 
-        echo json_encode($results);
+        echo json_encode( $results );
 
         die();
     }
@@ -192,18 +203,27 @@ class Ajax_Social_Invite_Helper {
         $subject = $_POST['subject'];
         $contacts = $_POST['contacts'];
         $message = $_POST['message'];
+        $output = array();
 
         if ( ! empty( $accessToken ) && ! empty( $provider ) ) {
-            $output = array();
 
-            if ($provider == "twitter" || $provider == "linkedin") {
-                for ($i = 0; $i < count($contacts); $i++) {
+            if ( $provider == "twitter" || $provider == "linkedin" ) {
+                for ( $i = 0; $i < count( $contacts ); $i++ ) {
 
                     $to = $contacts[$i]['Id'];
                     $to_name = $contacts[$i]['Name'];
-                    $response = $loginRadiusObject->loginradius_send_message( $accessToken, $to, $subject, $message );
 
-                    if (isset($response->isPosted) && $response->isPosted != '') {
+                    if( $provider == "twitter" ) {
+                        $message = substr( $message, 0, 140 );
+                    }
+
+                    try {
+                        $response = $loginRadiusObject->loginradius_send_message( $accessToken, $to, $subject, $message ); 
+                    } catch ( LoginRadiusException $e ) {
+                        $response = null;
+                    }
+                    
+                    if ( isset( $response->isPosted ) && $response->isPosted != '' ) {
                         $isPosted = true;
                     } else {
                         $isPosted = false;
@@ -215,14 +235,14 @@ class Ajax_Social_Invite_Helper {
                     );
                 }
             } else {
-                for ($i = 0; $i < count($contacts); $i++) {
+                for ( $i = 0; $i < count( $contacts ); $i++ ) {
                     $to = $contacts[$i]['Id'];
                     $to_name = $contacts[$i]['Name'];
 
                     $from_name = $user->display_name;
                     $from_email = $user->user_email;
 
-                    if (isset($lr_social_invite_settings['enable_custom_email']) && $lr_social_invite_settings['enable_custom_email'] == '1') {
+                    if ( isset( $lr_social_invite_settings['enable_custom_email'] ) && $lr_social_invite_settings['enable_custom_email'] == '1' ) {
                         $from_name = $lr_social_invite_settings['email_name'];
                         $from_email = $lr_social_invite_settings['email_address'];
                     }
@@ -231,7 +251,7 @@ class Ajax_Social_Invite_Helper {
                     $headers = 'From: ' . $from_name . ' <' . $from_email . '>' . "\r\n" .
                             'Reply-To: ' . $from_email . "\r\n" .
                             'X-Mailer: PHP/' . phpversion();
-                    $response = wp_mail($to, $subject, $message, $headers);
+                    $response = wp_mail( $to, $subject, $message, $headers );
 
                     $output[] = array(
                         "isPosted" => $response,
@@ -241,8 +261,7 @@ class Ajax_Social_Invite_Helper {
             }
         }
 
-        echo json_encode($output);
-        die();
+        die( json_encode( $output ) );
     }
 
 }
